Moved 'json_converter_spec.rb' to root directory, prevents it being used as a plugin
Created layouts for index, department, public body.

These layouts can be used for all sorts of cool stuff!
- Data is embedded into them which can be used at compile time
- Content for these is generated directly from the json
- Anything from the json can be used

Next we have json_converter.rb
This generates a full tree of html pages automatically based on templates. Before we were doing this in parse.py but it was misplaced.

parse.py has been greatly trimmed down. Anything to do with html is gone! We pump out json which is then processed into html.

This gives us the api, around which we can generate the user facing things